third_party/blink/web_tests/external/wpt/dom/traversal/TreeWalker-acceptNode-filter-cross-realm-expected.txt

This is a testharness.js-based test.
PASS parent's |this|, parent's method, parent's filter, parent's root
PASS parent's |this|, parent's method, parent's filter, iframe's root
FAIL parent's |this|, parent's method, iframe's filter, iframe's root assert_equals: expected function "function TypeError() { [native code] }" but got function "function TypeError() { [native code] }"
PASS parent's |this|, iframes's method, iframe's filter, iframe's root
Harness: the test ran to completion.
